refactor(processing/runner): reorganise - merging some runner and processing, and moving some elsewhere (#171)
