+++ tests/test_processing.py
@@ -8,69 +8,67 @@ from unittest.mock import patch
 
 import pytest
 
-from lintquarto.processing import (
-    gather_qmd_files,
-    process_qmd,
-    validate_no_commas,
-)
+from lintquarto.gather import gather_qmd_files
+from lintquarto.main import validate_no_commas
+from lintquarto.runner import lint_qmd
 
 CORE_LINTER = "flake8"
 
 
 # =============================================================================
-# 1. process_qmd()
+# 1. lint_qmd()
 # =============================================================================
 
 
-def test_process_qmd_with_real_file(tmp_path):
-    """Integration Test: process_qmd runs on a real .qmd file."""
+def test_lint_qmd_with_real_file(tmp_path):
+    """Integration Test: lint_qmd runs on a real .qmd file."""
     # Create a temporary quarto file
     qmd_file = tmp_path / "test.qmd"
     qmd_file.write_text("# Test Quarto file\n``````")
 
-    # Call process_qmd and attempt to lint it - should return a valid exit code
-    result = process_qmd(str(qmd_file), CORE_LINTER)
+    # Call lint_qmd and attempt to lint it - should return a valid exit code
+    result = lint_qmd(str(qmd_file), CORE_LINTER)
     assert result in (0, 1)
 
 
-def test_process_qmd_invalid_file(tmp_path):
-    """Integration Test: process_qmd returns error for invalid file."""
+def test_lint_qmd_invalid_file(tmp_path):
+    """Integration Test: lint_qmd returns error for invalid file."""
     # Run on a file that doesn't exist
-    result = process_qmd(str(tmp_path / "notfound.qmd"), CORE_LINTER)
+    result = lint_qmd(str(tmp_path / "notfound.qmd"), CORE_LINTER)
     assert result == 1
 
-    # Create a text file and attempt to run process_qmd()
+    # Create a text file and attempt to run lint_qmd()
     txt_file = tmp_path / "file.txt"
     txt_file.write_text("print('hello')")
-    result = process_qmd(str(txt_file), CORE_LINTER)
+    result = lint_qmd(str(txt_file), CORE_LINTER)
     assert result == 1
 
 
-def test_process_qmd_keep_temp(tmp_path):
-    """Integration Test: process_qmd keeps the temporary .py file."""
+def test_lint_qmd_keep_temp(tmp_path):
+    """Integration Test: lint_qmd keeps the temporary .py file."""
     # Create a temporary quarto file
     qmd_file = tmp_path / "test.qmd"
     qmd_file.write_text("# Test Quarto file\n``````")
 
-    # Call process_qmd and attempt to lint it
-    _ = process_qmd(str(qmd_file), CORE_LINTER, keep_temp_files=True)
+    # Call lint_qmd and attempt to lint it
+    _ = lint_qmd(str(qmd_file), CORE_LINTER, keep_temp_files=True)
 
-    # Assert that the .py file still exists after process_qmd returns
+    # Assert that the .py file still exists after lint_qmd returns
     py_file = tmp_path / "test.py"
     assert py_file.exists(), (
         "Temporary .py file should be kept when keep_temp_files=True"
     )
 
 
-def test_process_qmd_pylint_filepath(capsys):
+def test_lint_qmd_pylint_filepath(capsys):
     """Checks filepath in pylint output is not repeating folder names."""
     # Get path to the example QMD file that already produces pylint warnings.
     # This ensures we will have some output to check.
     test_dir = Path(__file__).parent
     qmd_path = test_dir / "examples" / "general_example.qmd"
 
-    # Run process_qmd with pylint on the example file and capture output.
-    _ = process_qmd(qmd_path, "pylint", keep_temp_files=False, verbose=True)
+    # Run lint_qmd with pylint on the example file and capture output.
+    _ = lint_qmd(qmd_path, "pylint", keep_temp_files=False, verbose=True)
     output = capsys.readouterr().out
 
     # Use regex to extract every filename prefix used in diagnostic lines.
@@ -181,7 +179,7 @@ def test_temp_file_cleaned_up_on_success(tmp_path):
     qmd_file = tmp_path / "test.qmd"
     qmd_file.write_text("```{python}\nx = 1\n```\n")
 
-    process_qmd(qmd_file, linter="flake8")
+    lint_qmd(qmd_file, linter="flake8")
 
     assert not any(tmp_path.glob("*.py"))
 
@@ -192,10 +190,10 @@ def test_temp_file_cleaned_up_on_exception(tmp_path):
     qmd_file.write_text("```{python}\nx = 1\n```\n")
 
     with patch(
-        "lintquarto.processing.subprocess.run",
+        "lintquarto.runner.subprocess.run",
         side_effect=RuntimeError("boom"),
     ):
-        ret = process_qmd(qmd_file, linter="flake8")
+        ret = lint_qmd(qmd_file, linter="flake8")
 
     assert ret == 1
     assert not any(tmp_path.glob("*.py"))
@@ -206,6 +204,6 @@ def test_temp_file_kept_when_flag_set(tmp_path):
     qmd_file = tmp_path / "test.qmd"
     qmd_file.write_text("```{python}\nx = 1\n```\n")
 
-    process_qmd(qmd_file, linter="flake8", keep_temp_files=True)
+    lint_qmd(qmd_file, linter="flake8", keep_temp_files=True)
 
     assert len(list(tmp_path.glob("*.py"))) == 1
